Fix mapping for g:Todo_txt_prefix_creation_date variable.
Update version 0.8.2->0.9.0

Before was checked only the existence of the g:Todo_txt_prefix_creation_date variable.
So the date would have been inserted even if the variable had been set to zero in .vimrc

In addition the mapping conflicts with the vim completion, if the selected item was inserted usin Carriage return and not the vim-ish <CTR-e> keys
In that case the completion item (for example the project or the context) would be followed by the date.