Disable broken test cases
Tomorrow's day on month is NOT today's day of month + 1. This test case
design assumes months are infinitely long, while in fact the day wrap
back to 1 after the 28:th, 29:th, 30:th or 31:th. Doh!

Leaving the code around, but commented out. In case someone with to
revive these test cases, some other cause of truth for the oracle value
needs to be used. There's not much point in calculating it in the same
way as the plugin. Frankly, one can also argue there isn't much point
in having these two test cases at all.